refactoring chakra providers in providers.tsx, to wrap around just the pages and components that need it, as opposed to around the {children} in the root template.tsx
