+++ frontend/src/app/(signup)/login/page.tsx
@@ -1,5 +1,6 @@
 import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
 import { cookies } from "next/headers";
+import { Providers } from "@/providers";
 
 import type { Database } from "../../../lib/database.types";
 import LoginForm from "../../../components/auth/LoginForm";
@@ -7,7 +8,9 @@ import LoginForm from "../../../components/auth/LoginForm";
 export default async function LoginPage() {
   return (
     <div>
-      <LoginForm />
+      <Providers>
+        <LoginForm />
+      </Providers>
     </div>
   );
 }
