+++ frontend/src/app/layout.tsx
@@ -1,7 +1,6 @@
 import "./globals.css";
 
 import { Inter } from "next/font/google";
-import { Providers } from "./providers";
 import Header from "../components/Header";
 import Footer from "../components/Footer";
 import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
@@ -26,9 +25,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
 
   return (
     <html lang="en">
-      <body className={inter.className}>
-        <Providers>{children}</Providers>
-      </body>
+      <body className={inter.className}>{children}</body>
     </html>
   );
 }
